+++ test/test_lexer_helper.py
@@ -336,3 +336,7 @@ class TestLexerHelper(unittest.TestCase):
         for filename in filenames:
             lexer = guess_lexer_wrapper(filename=filename, text='')
             self.assertEqual(lexer.pygments_lexer_cls, TextWithCommentLexer)
+
+    def test_npmrc_files(self):
+        lexer = guess_lexer_wrapper(filename='.npmrc', text='')
+        self.assertEqual(lexer.pygments_lexer_cls, IniLexer)