+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Add lexer support for `.npmrc` files.
+
 ## 1.1.2 - 2026-06-15
 
 * Add `-v, --version` flag to display the current version of `affe` and exit.